spritepal/requirements.txt
# SpritePal Requirements File
# Aligned with pyproject.toml

# Core Qt Framework (PySide6 - matches pyproject.toml)
PySide6>=6.5.0

# Image Processing
Pillow>=9.0.0
numpy>=1.21.0

# System Monitoring
psutil>=7.1.3

# Testing Framework
pytest>=7.0.0
pytest-qt>=4.2.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-timeout>=2.0.0

# Code Quality Tools
ruff>=0.1.0
basedpyright>=1.8.0
